+++ src/datatypes/index.ts
@@ -3,6 +3,7 @@ export type * from "@/datatypes/account.js";
 export type * from "@/datatypes/activity.js";
 export type * from "@/datatypes/analytics.js";
 export type * from "@/datatypes/api-token.js";
+export type * from "@/datatypes/artifact.js";
 export type * from "@/datatypes/audit.js";
 export type * from "@/datatypes/auth.js";
 export type * from "@/datatypes/catalog.js";